task4/pyproject.toml
[project]
name = "task4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf",
    "streamlit",
]
